web components generator diverging
