+++ .gitignore
@@ -1,3 +1,6 @@
-node_modules/
-temp/
+test/temp/
+/npm-debug.log
 /.idea
+bower_components/
+node_modules/
+/test/temp/my-element.html
