+++ README.md
@@ -1,85 +1,72 @@
-# Yeoman Generator<br> for Custom Elements<br>
-
-[![NPM version](http://img.shields.io/npm/v/generator-element.svg?style=flat)](http://npmjs.org/generator-element)
-[![NPM downloads](http://img.shields.io/npm/dm/generator-element.svg?style=flat)](http://npmjs.org/generator-element)
-[![Build Status](http://img.shields.io/travis/webcomponents/generator-element/master.svg?style=flat)](https://travis-ci.org/webcomponents/generator-element)
-[![Dependency Status](http://img.shields.io/david/webcomponents/generator-element.svg?style=flat)](https://david-dm.org/webcomponents/generator-element)
-
-![WC + Yeoman](https://cloud.githubusercontent.com/assets/398893/3528278/c3c2eae6-078d-11e4-88c5-7a853d7a5757.jpg)
+<h1 align="center">Yeoman Generator Webcomponent</h1>
+<p align="center">
+  <img title="ci alarm" src='https://github.com/yeoman/media/blob/master/optimized/yeoman-150x150-opaque.png' />
+</p>
+<p align="center">
+  <a title='Build Status' href="https://travis-ci.org/eromano/webcomponent-generator-element">
+    <img src='https://travis-ci.org/eromano/webcomponent-generator-element.svg?branch=master' />
+  </a>
+  <a href='https://coveralls.io/r/eromano/webcomponent-generator-element'>
+    <img src='https://img.shields.io/coveralls/eromano/webcomponent-generator-element.svg' alt='Coverage Status' />
+  </a>
+  <a href='https://github.com/eromano/ci-alarm/blob/master/LICENSE'>
+    <img src='https://img.shields.io/badge/license-MIT-blue.svg' alt='license' />
+  </a>
+</p>
 
 > A Yeoman Generator that provides a functional boilerplate to easily create Custom Elements using [Polymer](http://www.polymer-project.org/), [X-Tag](http://x-tags.org/) or [VanillaJS](http://vanilla-js.com/).
-
-> All templates are based in the boilerplates authored by the [WebComponents.org team](https://github.com/webcomponents/):
-
-> * [Polymer Boilerplate](https://github.com/webcomponents/polymer-boilerplate)
-> * [X-Tag Boilerplate](https://github.com/webcomponents/x-tag-boilerplate)
-> * [VanillaJS Boilerplate](https://github.com/webcomponents/element-boilerplate)
 
 ## Install
 
 Install this generator using NPM:
 
 ```sh
-$ [sudo] npm install -g generator-element
+$ [sudo] npm install -g webcomponent-generator-element
 ```
 
 ## Getting Started
 
-![ScreenShot](https://cloud.githubusercontent.com/assets/398893/3528261/a8bdefb6-078d-11e4-900d-b6d4df3ead79.png)
+```
+    $ mkdir name-new-webcomponent
+```
 
-There are two different generators available.
+```
+    $ cd name-new-webcomponent    
+```
 
-* The first one used to scaffold out new **individual elements**:
+``` 
+    $ yo webcomponent-generator-element    
+```
 
-    ```sh
-$ yo element
-    ```
+``` 
+    1. What do you want to use?
+    2. Which version?
+    3. What's the GitHub repository?
+    4. What's your GitHub username?
+    5. What's the name of your element?
+    6. How would you describe the element?
+    7. Do you want to include lifecycle callbacks?
+    8. Do you want to include some useful Grunt tasks?
+```
 
-    ```
-[?] What do you want to use?
-[?] What's the name of your element?
-[?] Do you want to include lifecycle callbacks?
-    ```
+Which will generate the following project structure:
 
-    Which will generate the following file:
 
-    ```
-.
-└── my-element.html
-    ```
 
-* The second one is used to scaffold an **entire project**:
+    ├── .editorconfig
+    ├── .gitignore
+    ├── bower.json
+    ├── package.json
+    ├── Gruntfile.js
+    ├── demo/
+    ├── test/
+    └── src/my-element.html
 
-    ```sh
-$ yo element:repo
-    ```
 
-    ```
-[?] What do you want to use?
-[?] What's the GitHub repository?
-[?] What's your GitHub username?
-[?] What's the name of your element?
-[?] How would you describe the element?
-[?] Do you want to include lifecycle callbacks?
-[?] Do you want to include some useful Grunt tasks?
-    ```
 
-    Which will generate the following project structure:
+    
+And run `bower install` & `npm install` for you to fetch all dependencies.
 
-    ```
-.
-├── .editorconfig
-├── .gitignore
-├── bower.json
-├── package.json
-├── index.html
-├── Gruntfile.js
-└── src/my-element.html
-    ```
-
-    And run `bower install` & `npm install` for you to fetch all dependencies.
-
-> _**Note**: files will be generated in the current directory, so be sure to change to a new directory before running those commands if you don't want to overwrite existing files._
 
 ## Contributing
 
@@ -91,16 +78,14 @@ $ yo element:repo
 
 ## History
 
-For detailed changelog, see [Releases](https://github.com/webcomponents/generator-element/releases).
+For detailed changelog, see [Releases](https://github.com/eromano/webcomponent-generator-element/releases).
+This project is a fork updated from [webcomponents/generator-element](https://github.com/webcomponents/generator-element).
+Contributors[contributors](https://github.com/eromano/webcomponent-generator-element/graphs/contributors).
 
-## Team
+## Contributors
 
-This project is maintained by these people and a bunch of awesome [contributors](https://github.com/webcomponents/generator-element/graphs/contributors).
-
-[![Zeno Rocha](https://2.gravatar.com/avatar/e190023b66e2b8aa73a842b106920c93)](https://github.com/zenorocha) | [![Addy Rocha](https://2.gravatar.com/avatar/96270e4c3e5e9806cf7245475c00b275)](https://github.com/addyosmani)
---- | --- | --- | --- | ---
-[Zeno Rocha](https://github.com/zenorocha) | [Addy Osmani](https://github.com/addyosmani)
-
-## License
-
-[MIT License](http://webcomponentsorg.mit-license.org/) © WebComponents.org
+Contributor | GitHub profile | Twitter profile |
+--- | --- | ---
+Eugenio Romano (contributor)| [Eugenio Romano](https://github.com/eromano) | [@RomanoEugenio](https://twitter.com/RomanoEugenio)
+Zeno Rocha  (creator) | [Zeno Rocha](https://github.com/zenorocha)| [@zenorocha](https://twitter.com/zenorocha)
+Addy Osmani (creator) | [Addy Osmani](https://github.com/addyosmani)| [@addyosmani](https://twitter.com/addyosmani)
