+++ banner.js
@@ -6,7 +6,6 @@ var banner = "" +
     chalk.green(" / / ")   + " | | | |/ _ \\ _____ / _ \\ |/ _ \\ '_ ` _ \\ / _ \\ '_ \\| __|  " + chalk.green("\\ \\ \n") +
     chalk.green(" \\ \\")  + "  | |_| | (_) |_____|  __/ |  __/ | | | | |  __/ | | | |_   "        + chalk.green("/ / \n") +
     chalk.green("  \\_\\") + "  \\__, |\\___/       \\___|_|\\___|_| |_| |_|\\___|_| |_|\\__| "   + chalk.green("/_/  \n") +
-                               "       |___/                                                        \n" +
-    chalk.green("\n    Custom Element generator brought to you by WebComponents.org\n");
+                               "       |___/                                                        \n" ;
 
 module.exports = banner;
