+++ package.json
@@ -1,13 +1,13 @@
 {
-  "name": "generator-element",
+  "name": "webcomponent-generator-element",
   "version": "1.0.0",
   "description": "Yeoman generator to create Custom Elements using Polymer, X-Tag or VanillaJS",
   "author": {
-    "name": "WebComponents.org",
-    "url": "https://github.com/webcomponents"
+    "name": "Eugenio Romano",
+    "url": "https://github.com/eromano/generator-element"
   },
   "license": "MIT",
-  "repository": "webcomponents/generator-element",
+  "repository": "eromano/generator-element",
   "main": "elements/index.js",
   "engines": {
     "node": ">=0.10.0"
@@ -22,7 +22,8 @@
     "webcomponents",
     "polymer",
     "x-tag",
-    "vanillajs"
+    "vanillajs",
+    "bosonic"
   ],
   "dependencies": {
     "chalk": "^1.0.0",
