+++ templates/_bower.json
@@ -7,7 +7,8 @@
     "main": "dist/<%= elementName %>.html",<% } %>
     "keywords": [<% if (boilerplate == 'Polymer') { %>
         "polymer",<% } %><% if (boilerplate == 'X-Tag') { %>
-        "x-tag",<% } %>
+        "x-tag",<% } %><% if (boilerplate == 'VanillaJS') { %>
+        "VanillaJS",<% } %>
         "web-components"
     ],
     "ignore": [
@@ -15,11 +16,26 @@
         "node_modules",
         "bower_components"
     ],
-    "dependencies":{<% if (boilerplate == 'Polymer') { %><% if (polymerVersion == '1.4.0') { %>
-        "polymer": "Polymer/polymer#^1.4.0"<% } %><% if (polymerVersion == '0.5.1') { %>
-        "polymer": "Polymer/polymer#^0.5.1"<% } %><% } %><% if (boilerplate == 'X-Tag') { %>
-        "x-tag-core": "^1.0.0",
-        "webcomponentsjs": "^0.7.2"<% } %><% if (boilerplate == 'VanillaJS') { %>
-        "webcomponentsjs": "^0.7.2"<% } %>
+    "dependencies":{
+        <% if (boilerplate == 'Polymer') { %>
+            <% if (polymerVersion == '1.4.0') { %>
+            "polymer": "Polymer/polymer#^1.4.0"
+            <% } %>
+            <% if (polymerVersion == '0.5.1') { %>
+            "polymer": "Polymer/polymer#^0.5.1"
+            <% } %>
+        <% } %>
+        <% if (boilerplate == 'X-Tag') { %>
+              <% if (xtagVersion == '1.5.0') { %>
+                "x-tag-core": "^1.5.0",
+              <% } %>
+              <% if (xtagVersion == '1.0.0') { %>
+                "x-tag-core": "^1.0.0",
+              <% } %>
+        "webcomponentsjs": "^0.7.2"
+        <% } %>
+        <% if (boilerplate == 'VanillaJS') { %>
+        "webcomponentsjs": "^0.7.2"
+        <% } %>
     }
 }
